requirements.txt
genson==1.2.2
requests==2.28.1
feedparser==6.0.10
requests-ratelimiter==0.4.0
pandas==1.3.5
SQLAlchemy==1.4.39
psycopg2-binary==2.9.3
python-decouple==3.6
schedule==1.0.0
pyyaml==5.4.0
requests-cache==1.0.0b0
python-dotenv==0.21.0
backoff==1.8.0
loguru==0.6.0
pytest==7.2.0
xmltodict==0.13.0
SQLAlchemy-Utils==0.38.3
snowflake-sqlalchemy==1.4.3
